Call BaseTemplate#printTrail before closing the open <body> and <html> tags
Definitely needed in order to not break things like the debug toolbar but it makes sense in general.
I did notice while testing this that for whatever reason the CSS loading still seems to be utterly broken if you slap ?debug=true to the URL, but I guess that's totally unrelated.
CSS is properly served in "production" mode and that's really what matters, and after this patch, the debug toolbar shows up appropriately on dev boxes as well.

Change-Id: I5e9a273231d4a63b76ee236ad0acb71385b0b6d8